Revert "rdiscd: disable accept_ra for duration of daemon"
Since the link has to be up for rdiscd to work, accept_ra has to already
be configured properly, or there's a race condition.  Maybe at some point
I'll make rdiscd responsible for bringing the link up, but until then,
it's pointless for rdiscd to be responsible for setting accept_ra.

This reverts commit 06e249a46e885bd61e5cc2e7e6605096225a5bf0.